+++ src/Importers/Ccp30/Injectors/FigurasTransporte.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatCatalogosPopulate\Importers\Ccp30\Injectors;
+
+use PhpCfdi\SatCatalogosPopulate\AbstractCsvInjector;
+use PhpCfdi\SatCatalogosPopulate\Database\DataFields;
+use PhpCfdi\SatCatalogosPopulate\Database\DataTable;
+use PhpCfdi\SatCatalogosPopulate\Database\DateDataField;
+use PhpCfdi\SatCatalogosPopulate\Database\PaddingDataField;
+use PhpCfdi\SatCatalogosPopulate\Database\TextDataField;
+use PhpCfdi\SatCatalogosPopulate\Utils\CsvFile;
+use RuntimeException;
+
+class FigurasTransporte extends AbstractCsvInjector
+{
+    public function checkHeaders(CsvFile $csv): void
+    {
+        $csv->move(3);
+        $expected = [
+            'Clave figura transporte',
+            'Descripción',
+            'Fecha inicio de vigencia',
+            'Fecha fin de vigencia',
+        ];
+        $headers = $csv->readLine();
+
+        if ($expected !== $headers) {
+            throw new RuntimeException("The headers did not match on file {$this->sourceFile()}");
+        }
+
+        $csv->next();
+    }
+
+    public function dataTable(): DataTable
+    {
+        return new DataTable('ccp_30_figuras_transporte', new DataFields([
+            new PaddingDataField('id', '0', 2),
+            new TextDataField('texto'),
+            new DateDataField('vigencia_desde'),
+            new DateDataField('vigencia_hasta'),
+        ]));
+    }
+}